test(install): move the arborist fixture cases out of migrate.test.ts
The 57 arborist snapshot cases run one after another, two bun children each,
and made migrate.test.ts the longest file of the serial phase. They move to
npm-arborist.test.ts with their snapshot file renamed byte for byte, and the
helpers both files need move to migration-harness.ts. No test changes.
